refactor: model 패키지를 domain과 service로 분리
