+++ src/main/java/lotto/Application.java
@@ -1,7 +1,7 @@
 package lotto;
 
 import lotto.controller.LottoController;
-import lotto.model.LottoService;
+import lotto.service.LottoService;
 import lotto.view.InputView;
 import lotto.view.OutputView;
