+++ src/main/java/lotto/domain/BonusNumber.java
@@ -1,4 +1,4 @@
-package lotto.model;
+package lotto.domain;
 
 public class BonusNumber {
