+++ src/main/java/lotto/domain/PurchaseAmount.java
@@ -1,4 +1,4 @@
-package lotto.model;
+package lotto.domain;
 
 public class PurchaseAmount {
