+++ src/main/java/lotto/domain/WinningNumbers.java
@@ -1,4 +1,4 @@
-package lotto.model;
+package lotto.domain;
 
 import java.util.List;
