+++ src/test/java/lotto/domain/BonusNumberTest.java
@@ -1,4 +1,4 @@
-package lotto.model;
+package lotto.domain;
 
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
