+++ src/test/java/lotto/domain/WinningNumbersTest.java
@@ -1,4 +1,4 @@
-package lotto.model;
+package lotto.domain;
 
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
